Added Feeders Controler and Token Filter

+++ src/main/java/com/feedme/server/repositories/UsersRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface UsersRepository extends CrudRepository<User, Integer> {
+public interface UsersRepository extends CrudRepository<User, Long> {
     /**
      * Finds users by name and password to check if password is correct.
      *
